requirements.txt
selenium
psycopg2-binary
webdriver-manager
bs4